[Refactor] Permission Handler 함수 변경
