+++ app/src/main/java/com/and04/naturealbum/ui/component/PermissionDialogs.kt
@@ -1,12 +1,11 @@
-package com.and04.naturealbum.ui.home
+package com.and04.naturealbum.ui.component
 
 import androidx.compose.runtime.Composable
 import com.and04.naturealbum.R
-import com.and04.naturealbum.ui.component.MyDialog
 
 enum class PermissionDialogState {
     None,
-    Explain,
+    ExplainCamera,
     GoToSettings,
 }
 
@@ -18,7 +17,7 @@ fun PermissionDialogs(
     onGoToSettings: () -> Unit = {},
 ) {
     when (permissionDialogState) {
-        PermissionDialogState.Explain -> {
+        PermissionDialogState.ExplainCamera -> {
             MyDialog(
                 onConfirmation = {
                     onDismiss()
